Fix: Changed cast from int to float in square.py

+++ Calculator/square.py
@@ -1,2 +1,2 @@
 def square(a):
-    return int(a) ** 2
+    return float(a) ** 2
